pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sear"
version = "2.3.0"
description = "SEAR: Summarization-Enhanced Augmented Retrieval - Multi-corpus RAG with FAISS retrieval, boolean queries, SQL interface, and LLM summarization"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Guard8.ai Team"},
]
maintainers = [
    {name = "Guard8.ai Team"}
]
keywords = [
    "rag",
    "retrieval",
    "faiss",
    "semantic-search",
    "ollama",
    "llm",
    "embedding",
    "vector-search",
    "multi-corpus",
    "augmented-retrieval",
    "boolean-queries",
    "sql-interface"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Indexing",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "sqlparse>=0.5.0",
]

[project.optional-dependencies]
# GPU support (requires CUDA toolkit installed)
gpu = [
    "faiss-gpu>=1.7.4",
]

# Anthropic Claude support for higher quality generation
anthropic = [
    "anthropic>=0.39.0",
]

# Document conversion support (PDF, DOCX to markdown)
converter = [
    "pdfplumber>=0.10.0",
    "pdf2image>=1.16.0",
    "pytesseract>=0.3.10",
    "python-bidi>=0.4.2",
    "Pillow>=10.0.0",
    "python-docx>=1.0.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# All optional features
all = [
    "sear[gpu,anthropic,converter]",
]

[project.urls]
Homepage = "https://github.com/Guard8ai/sear"
Documentation = "https://github.com/Guard8ai/sear#readme"
Repository = "https://github.com/Guard8ai/sear"
Issues = "https://github.com/Guard8ai/sear/issues"

[project.scripts]
sear = "sear.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["sear", "doc_converter", "doc_converter.core", "doc_converter.converters", "doc_converter.cli"]

# Black code formatter configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | build
  | dist
)/
'''

# Ruff linter configuration
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=sear",
    "--cov-report=term-missing",
]

# MyPy type checker configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "faiss.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "anthropic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pdfplumber.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pdf2image.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytesseract.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bidi.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "docx.*"
ignore_missing_imports = true
